+++ lib/src/circle_percentage.dart
@@ -4,10 +4,10 @@ import 'package:percentages_with_animation/src/extensions/extensions.dart';
 
 /// The circle percentage class provide you a to draw a circle percentage with different customization
 class CirclePercentage extends StatefulWidget {
-  /// This parameter is required and it's the current percentage value, (currentPercentage <= maxPercentage)
+  /// This parameter is required and it's the current percentage value, (currentPercentage <= maxPercentage & currentPercentage >= 0.0)
   final double currentPercentage;
 
-  /// This parameter is required and it's the maximum percentage value, (currentPercentage <= maxPercentage)
+  /// This parameter is required and it's the maximum percentage value, (maxPercentage >= currentPercentage)
   final double maxPercentage;
 
   /// This parameter is the circle size for the percentage with default size 100
@@ -31,6 +31,9 @@ class CirclePercentage extends StatefulWidget {
   /// This parameter is the text style of the label for the percentage text
   final TextStyle centerTextStyle;
 
+  /// This parameter is the call back to get the current percentage value during the animation (optional)
+  final Function(double)? onCurrentValue;
+
   const CirclePercentage({
     super.key,
     required this.currentPercentage,
@@ -42,6 +45,7 @@ class CirclePercentage extends StatefulWidget {
     this.backgroundColor = Colors.black12,
     this.centerText,
     this.centerTextStyle = const TextStyle(color: Colors.black),
+    this.onCurrentValue,
   })  : assert(currentPercentage <= maxPercentage),
         assert(currentPercentage >= 0),
         assert(duration >= 0);
@@ -62,6 +66,10 @@ class _CirclePercentageState extends State<CirclePercentage> {
       ),
       builder: (context, value, _) {
         double valueToShowOnText = value * widget.maxPercentage;
+        if (widget.onCurrentValue != null) {
+          widget.onCurrentValue!(
+              double.parse(valueToShowOnText.toStringAsFixed(2)));
+        }
         return Container(
           width: widget.size,
           height: widget.size,