+++ lib/src/gradient_circle_percentage.dart
@@ -4,10 +4,10 @@ import 'package:percentages_with_animation/src/extensions/extensions.dart';
 
 /// The Gradient percentage class provide you a to draw a circle gradient percentage with different customization
 class GradientCirclePercentage extends StatefulWidget {
-  /// This parameter is required and it's the current percentage value, (currentPercentage <= maxPercentage)
+  /// This parameter is required and it's the current percentage value, (currentPercentage <= maxPercentage & currentPercentage >= 0.0)
   final double currentPercentage;
 
-  /// This parameter is required and it's the maximum percentage value, (currentPercentage <= maxPercentage)
+  /// This parameter is required and it's the maximum percentage value, (maxPercentage >= currentPercentage)
   final double maxPercentage;
 
   /// This parameter is the circle size for the percentage with default size 100
@@ -37,6 +37,9 @@ class GradientCirclePercentage extends StatefulWidget {
   /// This parameter is the text style of the label for the percentage text
   final TextStyle centerTextStyle;
 
+  /// This parameter is the call back to get the current percentage value during the animation (optional)
+  final Function(double)? onCurrentValue;
+
   const GradientCirclePercentage({
     super.key,
     required this.currentPercentage,
@@ -50,6 +53,7 @@ class GradientCirclePercentage extends StatefulWidget {
     this.backgroundColor = Colors.white,
     this.centerText,
     this.centerTextStyle = const TextStyle(color: Colors.black),
+    this.onCurrentValue,
   })  : assert(currentPercentage <= maxPercentage),
         assert(currentPercentage >= 0),
         assert(duration >= 0);
@@ -71,6 +75,10 @@ class _GradientCirclePercentageState extends State<GradientCirclePercentage> {
       ),
       builder: (context, value, _) {
         double valueToShowOnText = value * widget.maxPercentage;
+        if (widget.onCurrentValue != null) {
+          widget.onCurrentValue!(
+              double.parse(valueToShowOnText.toStringAsFixed(2)));
+        }
         return Container(
           width: widget.size,
           height: widget.size,